+++ server/package.json
@@ -44,6 +44,7 @@
     "prismic-javascript": "^1.1.3",
     "raven": "^1.2.0",
     "rimraf": "^2.6.1",
+    "rss": "^1.2.2",
     "superagent": "^3.3.1",
     "supertest": "^3.0.0",
     "svgo": "^0.7.1"
